Add a public_name attribute to each node
This public name can be used by the user to specify a DNS name that will
point to the the public IP of this node. This is needed since the
deployer might not want to rely on automatic DNS names added for this
host in the subdomain used for the cloud.

Part of https://bugzilla.novell.com/show_bug.cgi?id=777523